netsync: allow fetchHeaderBlocks and fetchUtreexoHeaderBlocks to be specified a peer
Both of the functions currently default to the syncPeer for fetching
blocks and utreexo headers but if we allow our peers to send over
headers instead of invs, then we may not have a syncPeer as we're no
longer in ibd.

To allow these functions to be re-used when our peer sends us block
headers, we allow the caller to pass in the desired peer to request the
additional data from.