netsync: add numLeave, queuedBlocks and queuedUtreexoProofs to SyncManager
Because we need to have both the block and the utreexo proof in order to
validate a block as a CSN, we must be able to hold onto the proof or the
block until the other also arrives from the peer.

Having the queue maps allow us to do that.

The numLeaves allows us to keep track of leaves in memory from the
utreexo headers that we received. We need this to construct the get
utreexo proof message.