blockchain: add chain selection logic to maybeAcceptBlockHeader
maybeAcceptBlockHeader was missing the chain selection logic that would
be needed for handling block annoucements for ones that don't extend the
current chain tip.

It's now able to determine if a block header extends the chain tip or
not and returns a boolean indicating if the given header is on the
best chain of headers.