indexers: handle targets on flatutreexoproofindex init
On initialization, we use the targets from the targets flat file state
for turning the node from an archive to a pruned node. This is because
we're gonna store the targets in just the targets flat file state in
later commits. We also remove it completely as there's no need for a
pruned bridge to keep a part of the proof.