indexers, chaincfg, main: have multiple blockttl states
BlockTTL state is responsible for serving proofs to the peers for the
ttl messages. Since the committed TTLs are checkpointed at a given
height, we need to have multiple pollards so that we're able to provide
proofs for older clients when they ask for proofs for the older committed
states.