netsync: change getdatasize length to MaxUtreexoTTLPerMsg
During fetchHeaderBlocks, we'd allocate 50,000 (max invs per getdata)
inventory vectors in the getdata message which contributed
significantly to the memory allocated during ibd as we were
overallocating when downloading ttls first.

Since we can only download blocks up to the ttls that we have, we'll
only be getting 16 (MaxUtreexoTTLPerMsg) invs at a time while
allocating 50,000 invs.

Changing this here reduces the memory allocation and thus speeds up the
entire program.