aggregator: add helper functions for aggregator
For swiftsync, you must have an aggregator to check that all the utxos
marked as spent were indeed spent. We create a helper function to
efficiently add and subtract from the aggregator.

We can use big.Int but converting [32]byte hashes to big.Int values
allocate memory and slow down the ibd.